Hide plumbing (count{32,64}) behind an interface seqCounter
type seqCounter interface {
    countOf(seqString) counter
    allCounts(length int) []seqCount
}

All plumbing related to seq32/seq64 in frequencyReport and
sequenceReport is now hidden in the implementations
of this interface: seqCounts32, seqCounts64.

There is a cost to use calls through interfaces, but it is very limited
because those calls are not in hot loops.
And there is a huge benefit: the code is much more idiomatic.