Use an array instead of map for length <= 6
For length of 6, the number of different strings is 4096. This is small
enough to use an array for counting instead of a map.
This gives a HUGE performance boost: 11s instead of 15.5s.

Note that this may out of scope for the k-nucleotide challenge whose
puspose to to compare hash maps implementations.